Inventory UI, other UI
Bar3 is just a white cube. yes i know its just a white cube.

The side bar's width has been reduced from width/5 or 256 to 240. This is to match the sidebar from 7.0, also the camera has been changed in compensation to align with the new tab.

The player now recognizes "Expirience-Goal" and will display the stat when you are below level 20 on the exp bar.

Cleaned up the CharacterBars.cs file

And then just worked on the side bar inventory ui.

**AND** The hpSlot & mpSlot will be revised and the hotbar tabs to see stats screen/backpack haven't been added yet, this is just a visual change