UseItem, Backpack, Tabs, StatsView, Added new BoldText type
Added a basic useitem that just sends the packet if you try to use a consumable or item with type 10.

Added backpack ui & support
Added stat-view ui ( not finished )
Added the tabs on your inventory

Added a 3rd bold text, this was added because i want the flexibility of having a bolder text. because of how things where set up ive changed some types around and now:

0 = No bold
1 = Bold
2 = Bolder